feat(analyzer): make strict list index checks optional
In version `1.0.0-beta.30`, the check for list array access was changed to require a `non-negative-int` index type. While technically more correct, as negative indices are invalid for lists, this proved to be too restrictive for many common use cases.

This commit reverts the default behavior to allow any `int` as a list index.

To retain the benefits of the stricter validation, a new configuration option, `strict-list-index-checks`, has been introduced. When set to `true`, it enforces the non-negative integer requirement. This setting defaults to `false`, providing a more flexible out-of-the-box experience while allowing users to opt into stricter checks.

closes #565

Signed-off-by: azjezz <azjezz@protonmail.com>