fix(analyzer): Handle multiple docblocks before a statements
Currently we're only handling docblocks that are tied to a statement,
and only consider a single docblock to be tied to any given statement.
This means that having multiple consecutive docblocks to declare
variables, for example in a classic PHP-only template file, will have us
only considering the last of these docblocks. We can improve this by
treating all consecutive docblocks that preceed a given statement, only
separated by whitespace or comments, to apply to that statement.

This uncovered a bug in the docblock_var_on_non_assignment testcase
which actually didn't test what it was supposed to test because the
extra docblock was ignored. With the new behaviour, that test actually
complains that the docblock is redundant because it doesn't narrow the
known type. We can widen the original type and add a call that actually
requires the type-narrowing docblock to be processed.

We can also adjust the test case for issue #801 because we now fully
support the original code.